The return of SessionPlus
Added platform specific API tha rely of different packages.
